separating jinet test module
