+++ aydin/nn/models/torch/test/test_torch_unet.py
@@ -1,28 +1,13 @@
 # flake8: noqa
 
-# import math
-# from collections import OrderedDict
-# from itertools import chain
-#
-# import napari
 import numpy
-import pytest
 import torch
 
-# from torch.optim.lr_scheduler import ReduceLROnPlateau
-from torch.utils.data import DataLoader
-
 from aydin.io.datasets import add_noise, camera, normalise
-from aydin.nn.models.torch.torch_jinet import JINetModel
 from aydin.nn.models.torch.torch_unet import UNetModel, n2t_unet_train_loop
-from aydin.nn.models.utils.torch_dataset import TorchDataset
-
-# from aydin.nn.pytorch.it_ptcnn import to_numpy
-# from aydin.nn.pytorch.optimizers.esadam import ESAdam
-# from aydin.util.log.log import lprint
 
 
-def test_forward_2D_unet():
+def test_forward_2D():
     input_array = torch.zeros((1, 1, 64, 64))
     model2d = UNetModel(
         # (64, 64, 1),
@@ -62,7 +47,7 @@ def test_supervised_2D_n2t():
     # assert result.dtype == input_image.dtype
 
 
-def test_masking_2D_unet():
+def test_masking_2D():
     input_array = torch.zeros((1, 1, 64, 64))
     model2d = UNetModel(
         # (64, 64, 1),
@@ -75,15 +60,7 @@ def test_masking_2D_unet():
     assert result.dtype == input_array.dtype
 
 
-def test_forward_2D_jinet():
-    input_array = torch.zeros((1, 1, 64, 64))
-    model2d = JINetModel((64, 64, 1), spacetime_ndim=2)
-    result = model2d.predict([input_array])
-    assert result.shape == input_array.shape
-    assert result.dtype == input_array.dtype
-
-
-def test_supervised_3D_unet():
+def test_supervised_3D():
     input_array = torch.zeros((1, 1, 64, 64, 64))
     model3d = UNetModel(
         # (64, 64, 64, 1),
@@ -96,7 +73,7 @@ def test_supervised_3D_unet():
     assert result.dtype == input_array.dtype
 
 
-def test_masking_3D_unet():
+def test_masking_3D():
     input_array = torch.zeros((1, 1, 64, 64, 64))
     model3d = UNetModel(
         # (64, 64, 64, 1),
